theme2_health_v2.4/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
supabase>=2.0.0
psycopg2-binary>=2.9.0
google-api-python-client>=2.70.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-generativeai>=0.3.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0
pydantic>=2.0.0
websockets>=12.0
asyncio>=3.4.3
httpx[http2]>=0.25.0
realtime>=1.0.0
postgrest>=0.10.0
gotrue>=2.0.0
gspread>=5.0.0
google-auth-oauthlib>=0.8.0